W-12687160: Avoid keeping document reference in memory
